Revert "Shifted invocation logic into a new trait"
This reverts commit 9d378b9596ac908336ebf00fb66ab17079f07536.

It is necessary because, when creating an `InvocationFailedException`,
the new instance must receive the instance of the invoker, but that is
not possible if the invocation happens in a trait, because it is not
a class that implements `InvokerInterface`. This was not visible before
because the current exception factory is missing the `$invoker` param.